reduce size and improve readme
