+++ src/index.ts
@@ -1,5 +1,1 @@
-import { FlagProvider } from "./context";
-import FlagGuard from "./guard";
-import FlagSwitch from "./switch";
-
-export { FlagGuard, FlagProvider, FlagSwitch };
+export * from './flag-poles'
